Further cleanup after conversion of signed/unsigned xxx int to YintXX_t variables. Use of Baud rate constant and its value from 9600 to 9600U. Typecasting inside voltage computations for ADCs.

+++ e13_console_keil/main.c
@@ -10,12 +10,12 @@
 
 #include <stdio.h>
 
-#define BAUDRATE     (9600)             // Baud rate of UART in bps
+#define BAUDRATE     9600U              // Baud rate of UART in bps
 
 void main(void)
 {
     unsigned char state;
-    unsigned int  counter = 0;
+    unsigned long int counter = 0;
     static xdata char buffer[4];
     static xdata char buf_test[20];
