+++ e28_reentrance/interrupt.c
@@ -14,7 +14,7 @@
 void Timer4_Init (uint32_t sysclock, uint32_t rate)
 {
     uint8_t SFRPAGE_SAVE = SFRPAGE;        // Save Current SFR page
-    int32_t counts = sysclock/(12*rate);   // Note that Timer4 is connected to SYSCLK/12
+   uint16_t counts = (uint16_t)( sysclock/(12L*rate) ); // Note that Timer4 is connected to SYSCLK/12
 
     SFRPAGE = CONFIG_PAGE;              // set SFR page
     P3MDOUT |= 0x3F;                    // Set P3.0 through P3.5 to push-pull
@@ -23,8 +23,8 @@ void Timer4_Init (uint32_t sysclock, uint32_t rate)
     TMR4CN  = 0x00;                     // Stop Timer4; Clear TF4;
     TMR4CF  = 0x00;                     // use SYSCLK/12 as timebase
 //  TMR4CF  = 0x08;                     // use SYSCLK as timebase
-    RCAP4   = 65536 -(uint16_t)counts;  // Init reload values
-    // or   = -(uint16_t)counts; -- see the in class comment
+    RCAP4   = (uint16_t)(65536U - counts); // Init reload values
+    // or   = -counts; -- see the in class comment
     TMR4    = RCAP4;                    // set starting value
     EIE2   |= 0x04;                     // enable Timer4 interrupts - bit 00000100 or ET4 = 1;
     TMR4CN |= 0x04;                     // start Timer4
