+++ e33_measure_8ch_robin/main.c
@@ -11,8 +11,8 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#define BAUDRATE      (9600)            // Baud rate of UART in bps
-#define SAMPLE_RATE (10000L)            // Sample frequency in Hz
+#define BAUDRATE       9600U            // Baud rate of UART in bps
+#define SAMPLE_RATE   10000L            // Sample frequency in Hz
 
 void main(void)
 {
@@ -33,12 +33,12 @@ void main(void)
     while (1)
     {
         __xdata char buffer[16];        // character buffer for outputting one formatted value
-        int16_t ch;
+        uint8_t ch;
         for(ch=0; ch<8; ch++)
         {
             // Read the latest A/D results
             uint16_t value = getRecentResult(ch);
-            int16_t converted = (int32_t)value * VREF / ADC0_MAX;
+            uint16_t converted = (int32_t)value * (uint16_t)VREF / ADC0_MAX;
 
             sprintf(buffer, " ch%1d:%4d", ch, converted );  // note: !! formatting works only on xdata buffer !!
 
@@ -47,8 +47,8 @@ void main(void)
         }
 
         {
-            uint16_t temperature;   // temperature in hundredths of a degree C
-            int16_t temp_int, temp_frac;    // integer and fractional portions of temperature
+            uint16_t temperature;           // temperature in hundredths of a degree C
+            uint16_t temp_int, temp_frac;   // integer and fractional portions of temperature
 
             temperature = getRecentResult(8) <<1; // compensate for gain==1 instead of ==2
